shop/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
